+++ .babelrc
@@ -1,4 +1,13 @@
 {
-  "presets": [ "es2015-rollup" ],
-  "plugins": ["transform-class-properties"]
+  "presets": [
+    "stage-2",
+    ["env", {
+      "targets": {
+        "browsers": ["last 2 versions", "ie 11"]
+      }
+    }]
+  ],
+  "plugins": [
+    "transform-runtime"
+  ]
 }
