+++ .eslintrc
@@ -3,6 +3,7 @@ parser: babel-eslint
 extends: airbnb/base
 rules:
   import/no-extraneous-dependencies: off
+  consistent-return: off
   no-confusing-arrow: off
   no-console: off
   no-plusplus: off
