+++ .gitignore
@@ -1,3 +1,4 @@
+coverage
 node_modules
 npm-debug.log
 .DS_Store
