+++ package.json
@@ -14,33 +14,26 @@
     "build": "./node_modules/gulp/bin/gulp.js build",
     "build-watch": "./node_modules/gulp/bin/gulp.js build-watch",
     "lint": "eslint . --fix",
-    "test": "open http://localhost:8000/test/runner.html && python -m SimpleHTTPServer",
+    "test": "jest --coverage --testEnvironment=jsdom",
     "pages": "git push -f origin master:gh-pages",
     "postpublish": "git tag $npm_package_version && git push origin --tags"
   },
+  "dependencies": {
+    "babel-polyfill": "^6.23.0"
+  },
   "devDependencies": {
     "babel-eslint": "^7.2.3",
-    "babel-plugin-transform-class-properties": "^6.18.0",
-    "babel-preset-es2015": "6.9.0",
-    "babel-preset-es2015-rollup": "1.1.1",
-    "chai": "3.2.0",
+    "babel-jest": "^20.0.3",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.5.2",
+    "babel-preset-stage-2": "^6.24.1",
+    "chai": "^4.0.2",
     "eslint": "^3.19.0",
     "eslint-config-airbnb": "^13.0.0",
     "eslint-plugin-dollarshaveclub": "^1.0.2",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.10.3",
-    "gulp": "^3.9.0",
-    "gulp-header": "^1.8.8",
-    "jquery": "^2.1.4",
-    "mocha": "2.2.5",
-    "mocha-phantomjs": "3.6.0",
-    "phantomjs": "^2.1.7",
-    "rollup-plugin-babel": "2.6.1",
-    "rollup-plugin-uglify": "^1.0.1",
-    "rollup-stream": "^1.11.0",
-    "sinon": "1.15.4",
-    "uglify-js": "^2.7.4",
-    "vinyl-source-stream": "^1.1.0"
+    "jest": "^20.0.4"
   }
 }
