+++ src/__tests__/.eslintrc
@@ -1,0 +1,4 @@
+---
+env:
+  jasmine: true
+  jest: true
