--- test/.eslintrc
@@ -1,7 +1,0 @@
----
-env:
-  mocha: true
-globals:
-  chai: true
-rules:
-  no-unused-expressions: off
