--- test/runner.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Mocha Tests</title>
-    <link rel="stylesheet" href="../node_modules/mocha/mocha.css">
-    <meta charset="utf-8">
-    <script src="../node_modules/mocha/mocha.js"></script>
-    <script src="../node_modules/sinon/pkg/sinon-1.15.4.js"></script>
-    <script src="../node_modules/chai/chai.js"></script>
-    <script src="../build/study.min.js"></script>
-    <script src="./test1.js"></script>
-
-  </head>
-  <body>
-    <div id="mocha"></div>
-    <script>mocha.setup('bdd')</script>
-    <script src="./test2.js"></script>
-    <script>
-      if (window.mochaPhantomJS) { mochaPhantomJS.run(); }
-      else { mocha.run(); }
-    </script>
-  </body>
-</html>
